fix: resolve all ESLint errors and warnings failing CI
- Rename debug-user-links.ts → debugUserLinks.ts (camelCase filename)
- Rename verify-pin.post.ts → verifyPin.post.ts (camelCase filename)
- Add eslint-disable for process.env in playwright.config.ts (runs outside Nuxt)
- Prefix unused variables with _ (tokenExpiry, err, error)
- Import Buffer from node:buffer in picker.ts and thumbnail.get.ts
- Remove useless try/catch wrapper in getPickerItems
- Replace all console.log/error with consola across google-photos files and scripts
